c nested loops

+++ 0x02-functions_nested_loops/101-natural.c
@@ -4,7 +4,7 @@
  */
 int main()
 {
-long unsigned int i, sum = 0;
+int i, sum = 0;
 for (i = 0; i < 1024; i++)
 {
 if ((i % 5 == 0) || (i % 3 == 0))
